backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
gitpython
SQLAlchemy
alembic
psycopg2